Split operator expressions out to their own files
