+++ rule/expr.go
@@ -31,332 +31,6 @@ type Params interface {
 	EncodeValue(key string) (string, error)
 }
 
-type exprNot struct {
-	operator
-}
-
-func newExprNot() *exprNot {
-	return &exprNot{
-		operator: operator{
-			contract: Contract{
-				OpCode:     "not",
-				ReturnType: BOOLEAN,
-				Terms:      []Term{{Type: BOOLEAN, Cardinality: ONE}},
-			},
-		},
-	}
-}
-
-// Not creates an expression that evaluates the given operand e and returns its opposite.
-// e must evaluate to a boolean.
-func Not(e Expr) Expr {
-	expr := newExprNot()
-	expr.consumeOperands(e)
-	return expr
-}
-
-func (n *exprNot) Eval(params Params) (*Value, error) {
-	if len(n.operands) < 1 {
-		return nil, errors.New("invalid number of operands in Not func")
-	}
-
-	op := n.operands[0]
-	v, err := op.Eval(params)
-	if err != nil {
-		return nil, err
-	}
-
-	if v.Type != "bool" {
-		return nil, errors.New("invalid operand type for Not func")
-	}
-
-	if v.Equal(BoolValue(true)) {
-		return BoolValue(false), nil
-	}
-
-	return BoolValue(true), nil
-}
-
-type exprOr struct {
-	operator
-}
-
-func newExprOr() *exprOr {
-	return &exprOr{
-		operator: operator{
-			contract: Contract{
-				OpCode:     "or",
-				ReturnType: BOOLEAN,
-				Terms:      []Term{{Type: BOOLEAN, Cardinality: MANY, Min: 2}},
-			},
-		},
-	}
-}
-
-// Or creates an expression that takes at least two operands and evaluates to true if one of the operands evaluates to true.
-// All the given operands must evaluate to a boolean.
-func Or(vN ...Expr) Expr {
-	e := newExprOr()
-	e.consumeOperands(vN...)
-	return e
-}
-
-func (n *exprOr) Eval(params Params) (*Value, error) {
-	if len(n.operands) < 2 {
-		return nil, errors.New("invalid number of operands in Or func")
-	}
-
-	opA := n.operands[0]
-	vA, err := opA.Eval(params)
-	if err != nil {
-		return nil, err
-	}
-	if vA.Type != "bool" {
-		return nil, errors.New("invalid operand type for Or func")
-	}
-
-	if vA.Equal(BoolValue(true)) {
-		return vA, nil
-	}
-
-	for i := 1; i < len(n.operands); i++ {
-		vB, err := n.operands[i].Eval(params)
-		if err != nil {
-			return nil, err
-		}
-		if vB.Type != "bool" {
-			return nil, errors.New("invalid operand type for Or func")
-		}
-
-		if vB.Equal(BoolValue(true)) {
-			return vB, nil
-		}
-	}
-
-	return BoolValue(false), nil
-}
-
-type exprAnd struct {
-	operator
-}
-
-func newExprAnd() *exprAnd {
-	return &exprAnd{
-		operator: operator{
-			contract: Contract{
-				OpCode:     "and",
-				ReturnType: BOOLEAN,
-				Terms: []Term{
-					{
-						Type:        BOOLEAN,
-						Cardinality: MANY,
-						Min:         2,
-					},
-				},
-			},
-		},
-	}
-}
-
-// And creates an expression that takes at least two operands and evaluates to true if all the operands evaluate to true.
-// All the given operands must evaluate to a boolean.
-func And(vN ...Expr) Expr {
-	e := newExprAnd()
-	e.consumeOperands(vN...)
-	return e
-}
-
-func (n *exprAnd) Eval(params Params) (*Value, error) {
-	if len(n.operands) < 2 {
-		return nil, errors.New("invalid number of operands in And func")
-	}
-
-	opA := n.operands[0]
-	vA, err := opA.Eval(params)
-	if err != nil {
-		return nil, err
-	}
-	if vA.Type != "bool" {
-		return nil, errors.New("invalid operand type for And func")
-	}
-
-	if vA.Equal(BoolValue(false)) {
-		return vA, nil
-	}
-
-	for i := 1; i < len(n.operands); i++ {
-		vB, err := n.operands[i].Eval(params)
-		if err != nil {
-			return nil, err
-		}
-		if vB.Type != "bool" {
-			return nil, errors.New("invalid operand type for And func")
-		}
-
-		if vB.Equal(BoolValue(false)) {
-			return vB, nil
-		}
-	}
-
-	return BoolValue(true), nil
-}
-
-type exprEq struct {
-	operator
-}
-
-func newExprEq() *exprEq {
-	return &exprEq{
-		operator: operator{
-			contract: Contract{
-				OpCode:     "eq",
-				ReturnType: BOOLEAN,
-				Terms: []Term{
-					{
-						Type:        ANY,
-						Cardinality: MANY,
-						Min:         2,
-					},
-				},
-			},
-		},
-	}
-}
-
-// Eq creates an expression that takes at least two operands and evaluates to true if all the operands are equal.
-func Eq(vN ...Expr) Expr {
-	e := newExprEq()
-	e.consumeOperands(vN...)
-	return e
-}
-
-func (n *exprEq) Eval(params Params) (*Value, error) {
-	if len(n.operands) < 2 {
-		return nil, errors.New("invalid number of operands in Eq func")
-	}
-
-	opA := n.operands[0]
-	vA, err := opA.Eval(params)
-	if err != nil {
-		return nil, err
-	}
-
-	for i := 1; i < len(n.operands); i++ {
-		vB, err := n.operands[i].Eval(params)
-		if err != nil {
-			return nil, err
-		}
-
-		if !vA.Equal(vB) {
-			return BoolValue(false), nil
-		}
-	}
-
-	return BoolValue(true), nil
-}
-
-type exprIn struct {
-	operator
-}
-
-func newExprIn() *exprIn {
-	return &exprIn{
-		operator: operator{
-			contract: Contract{
-				OpCode:     "in",
-				ReturnType: BOOLEAN,
-				Terms: []Term{
-					{
-						Type:        ANY,
-						Cardinality: ONE,
-					},
-					{
-						Type:        ANY,
-						Cardinality: MANY,
-						Min:         1,
-					},
-				},
-			},
-		},
-	}
-}
-
-// In creates an expression that takes at least two operands and evaluates to true if the first one is equal to one of the others.
-func In(vN ...Expr) Expr {
-	e := newExprIn()
-	e.consumeOperands(vN...)
-	return e
-}
-
-func (n *exprIn) Eval(params Params) (*Value, error) {
-	if len(n.operands) < 2 {
-		return nil, errors.New("invalid number of operands in In func")
-	}
-
-	toFind := n.operands[0]
-	vA, err := toFind.Eval(params)
-	if err != nil {
-		return nil, err
-	}
-
-	for i := 1; i < len(n.operands); i++ {
-		vB, err := n.operands[i].Eval(params)
-		if err != nil {
-			return nil, err
-		}
-
-		if vA.Equal(vB) {
-			return BoolValue(true), nil
-		}
-	}
-
-	return BoolValue(false), nil
-}
-
-// exprIntToFloat is an operation that converts an Int64 type to a Float64 type.
-type exprIntToFloat struct {
-	operator
-}
-
-// newExprIntToFloat returns a pointer to a exprIntToFloat operator, initialised with a Contract.
-func newExprIntToFloat() *exprIntToFloat {
-	return &exprIntToFloat{
-		operator: operator{
-			contract: Contract{
-				OpCode:     "intToFloat",
-				ReturnType: FLOAT,
-				Terms: []Term{
-					{
-						Cardinality: ONE,
-						Type:        INTEGER,
-					},
-				},
-			},
-		},
-	}
-}
-
-// IntToFloat is an operator that converts as Float64 into an Int64 type.
-func IntToFloat(vN ...Expr) Expr {
-	e := newExprIntToFloat()
-	e.consumeOperands(vN...)
-	return e
-}
-
-// Eval will convert the operand provided to exprIntToFloat to a Float64Value.  Eval makes exprIntToFloat implement the Expr interface.
-func (n *exprIntToFloat) Eval(params Params) (*Value, error) {
-	op := n.operands[0]
-	val, err := op.Eval(params)
-	if err != nil {
-		return nil, err
-	}
-	iVal, err := strconv.ParseInt(val.Data, 10, 64)
-	if err != nil {
-		return nil, err
-	}
-	return Float64Value(float64(iVal)), nil
-}
-
 // Param is an expression used to select a parameter passed during evaluation and return its corresponding value.
 type Param struct {
 	Kind string `json:"kind"`
