+++ rule/expr_test.go
@@ -59,170 +59,6 @@ func (m *mockExpr) Contract() rule.Contract {
 	}
 }
 
-func TestNot(t *testing.T) {
-	t.Run("Eval/true", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(true), returnType: rule.BOOLEAN}
-		not := rule.Not(&m1)
-		val, err := not.Eval(nil)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(false), val)
-		require.Equal(t, 1, m1.evalCount)
-	})
-
-	t.Run("Eval/false", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(false)}
-		not := rule.Not(&m1)
-		val, err := not.Eval(nil)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(true), val)
-		require.Equal(t, 1, m1.evalCount)
-	})
-
-	t.Run("Eval/error", func(t *testing.T) {
-		m1 := mockExpr{val: rule.StringValue("foobar")}
-		not := rule.Not(&m1)
-		_, err := not.Eval(nil)
-		require.Error(t, err)
-		require.Equal(t, 1, m1.evalCount)
-	})
-}
-
-func TestAnd(t *testing.T) {
-	t.Run("Eval/true", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(true)}
-		m2 := mockExpr{val: rule.BoolValue(true)}
-		and := rule.And(&m1, &m2)
-		val, err := and.Eval(nil)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(true), val)
-		require.Equal(t, 1, m1.evalCount)
-		require.Equal(t, 1, m2.evalCount)
-	})
-
-	t.Run("Eval/short-circuit", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(false)}
-		m2 := mockExpr{val: rule.BoolValue(true)}
-		and := rule.And(&m1, &m2)
-		val, err := and.Eval(nil)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(false), val)
-		require.Equal(t, 1, m1.evalCount)
-		require.Equal(t, 0, m2.evalCount)
-	})
-
-	t.Run("Eval/false", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(true)}
-		m2 := mockExpr{val: rule.BoolValue(false)}
-		and := rule.And(&m1, &m2)
-		val, err := and.Eval(nil)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(false), val)
-		require.Equal(t, 1, m1.evalCount)
-		require.Equal(t, 1, m2.evalCount)
-	})
-
-	t.Run("Eval/error", func(t *testing.T) {
-		m1 := mockExpr{val: rule.StringValue("foobar")}
-		m2 := mockExpr{val: rule.BoolValue(true)}
-		and := rule.And(&m1, &m2)
-		_, err := and.Eval(nil)
-		require.Error(t, err)
-	})
-}
-
-func TestOr(t *testing.T) {
-	t.Run("Eval/true", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(true)}
-		m2 := mockExpr{val: rule.BoolValue(true)}
-		or := rule.Or(&m1, &m2)
-		val, err := or.Eval(nil)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(true), val)
-		require.Equal(t, 1, m1.evalCount)
-		require.Equal(t, 0, m2.evalCount)
-	})
-
-	t.Run("Eval/short-circuit", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(false)}
-		m2 := mockExpr{val: rule.BoolValue(true)}
-		or := rule.Or(&m1, &m2)
-		val, err := or.Eval(nil)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(true), val)
-		require.Equal(t, 1, m1.evalCount)
-		require.Equal(t, 1, m2.evalCount)
-	})
-
-	t.Run("Eval/false", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(false)}
-		m2 := mockExpr{val: rule.BoolValue(false)}
-		or := rule.Or(&m1, &m2)
-		val, err := or.Eval(nil)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(false), val)
-		require.Equal(t, 1, m1.evalCount)
-		require.Equal(t, 1, m2.evalCount)
-	})
-
-	t.Run("Eval/error", func(t *testing.T) {
-		m1 := mockExpr{val: rule.StringValue("foobar")}
-		m2 := mockExpr{val: rule.BoolValue(true)}
-		or := rule.Or(&m1, &m2)
-		_, err := or.Eval(nil)
-		require.Error(t, err)
-	})
-}
-
-func TestEq(t *testing.T) {
-	t.Run("Eval/Match", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(true)}
-		m2 := mockExpr{val: rule.BoolValue(true)}
-		params := regula.Params{"foo": "bar"}
-		eq := rule.Eq(&m1, &m2)
-		val, err := eq.Eval(params)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(true), val)
-		require.Equal(t, 1, m1.evalCount)
-		require.Equal(t, 1, m2.evalCount)
-		require.Equal(t, params, m1.lastParams)
-		require.Equal(t, params, m2.lastParams)
-	})
-
-	t.Run("Eval/NoMatch", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(true)}
-		m2 := mockExpr{val: rule.BoolValue(false)}
-		eq := rule.Eq(&m1, &m2)
-		val, err := eq.Eval(nil)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(false), val)
-	})
-}
-
-func TestIn(t *testing.T) {
-	t.Run("Eval/OK", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(true)}
-		m2 := mockExpr{val: rule.BoolValue(true)}
-		params := regula.Params{"foo": "bar"}
-		in := rule.In(&m1, &m2)
-		val, err := in.Eval(params)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(true), val)
-		require.Equal(t, 1, m1.evalCount)
-		require.Equal(t, 1, m2.evalCount)
-		require.Equal(t, params, m1.lastParams)
-		require.Equal(t, params, m2.lastParams)
-	})
-
-	t.Run("Eval/Fail", func(t *testing.T) {
-		m1 := mockExpr{val: rule.BoolValue(true)}
-		m2 := mockExpr{val: rule.BoolValue(false)}
-		eq := rule.In(&m1, &m2)
-		val, err := eq.Eval(nil)
-		require.NoError(t, err)
-		require.Equal(t, rule.BoolValue(false), val)
-	})
-}
-
 func TestParam(t *testing.T) {
 	t.Run("OK", func(t *testing.T) {
 		v := rule.StringParam("foo")
